// src/types/dashboard/dashboard-types.ts
export interface DashboardData {
    totalMatriculas: number;
    totalCursos: number;
    totalTurmas: number;
    matriculasAtivas: number;
    matriculasRecusadas: number;
    matriculasPendentes: number;
    matriculasCanceladas: number;
    turmasInativas: number;
    cursosInativos: number;
}